Checkered Board
Write a function which takes one parameter representing the dimensions of a checkered board. The board will always be square, 
so 5 means you will need a 5x5 board.

The dark squares will be represented by a unicode white square, while the light squares will be represented by a unicode black 
square (the opposite colours ensure the board doesn't look reversed on code wars' dark background). It should return a string of 
the board with a space in between each square and taking into account new lines.